+++ app/scripts/dimApp.i18n.js
@@ -175,7 +175,12 @@
             Consumables: "Consumables",
             Bounties: "Bounties",
             ShadersAndEmblems: "Shaders & Emblems",
-            Emotes: "Emotes" }
+            Emotes: "Emotes" },
+          TrialsCard: {
+            FiveWins: "5 Win Reward (Armor)",
+            SevenWins: "7 Win Reward (Weapon)",
+            Flawless: "Flawless"
+          }
         })
         .translations('it', {
           Level: "Livello",